feat: [1.2.0]の配信。

+++ Editor/GenericAssetsCreator.cs
@@ -10,12 +10,14 @@ namespace GDAT.Editor
     /// </summary>
     public class GenericAssetsCreator : UnityEditor.Editor
     {
-        const string MENU_ROOT = @"Tools/GameDevelopmentAuxiliaryToolsforUnity/";
+        private const string ASSETS_MENU_ROOT = @"Assets/Tools/GDATforUnity/";
+        private const string SPECIAL_FOLDERPATH = ASSETS_MENU_ROOT + @"Add Special Folders/";
+
         /// <summary>
         /// 選択フォルダ配下に汎用的なフォルダ(Animations,Audios等)を作成します。
         /// 既に存在するフォルダは変更しません。
         /// </summary>
-        [MenuItem(MENU_ROOT + "CreateGenericFolders", priority = 0)]
+        [MenuItem(ASSETS_MENU_ROOT + "Add Generic Folders", priority = 15)]
         private static void CreateGenericFoldersUnderSelectedFolders()
         {
             foreach (var asset in Selection.objects)
@@ -48,7 +50,7 @@ namespace GDAT.Editor
                         {
                             continue;
                         }
-                        FolderCreate(filePath);
+                        CreateDirectoryWithDummyFile(filePath);
                     }
 
                     AssetDatabase.Refresh();
@@ -56,29 +58,32 @@ namespace GDAT.Editor
             }
         }
 
-        private const string SPECIAL_FOLDERPATH = MENU_ROOT + @"SpecialFoleders/";
-
         #region 特別フォルダ生成メソッド
 
         /// <summary> 特別なフォルダを生成します </summary>
-        [MenuItem(SPECIAL_FOLDERPATH + "Resources", priority = 0)]
-        private static void ResourcesFolderCreate() => FolderCreate(@"Assets/Resources", true);
+        [MenuItem(SPECIAL_FOLDERPATH + "Resources", priority = 15)]
+        private static void ResourcesFolderCreate() => CreateDirectoryWithDummyFile(@"Assets/Resources", true);
         /// <summary> 特別なフォルダを生成します </summary>
-        [MenuItem(SPECIAL_FOLDERPATH + "Editor", priority = 0)]
-        private static void EditorFolderCreate() => FolderCreate(@"Assets/Editor", true);
+        [MenuItem(SPECIAL_FOLDERPATH + "Editor", priority = 15)]
+        private static void EditorFolderCreate() => CreateDirectoryWithDummyFile(@"Assets/Editor", true);
         /// <summary> 特別なフォルダを生成します </summary>
-        [MenuItem(SPECIAL_FOLDERPATH + "Gizmos", priority = 0)]
-        private static void GizmosFolderCreate() => FolderCreate(@"Assets/Gizmos", true);
+        [MenuItem(SPECIAL_FOLDERPATH + "Gizmos", priority = 15)]
+        private static void GizmosFolderCreate() => CreateDirectoryWithDummyFile(@"Assets/Gizmos", true);
         /// <summary> 特別なフォルダを生成します </summary>
-        [MenuItem(SPECIAL_FOLDERPATH + "Plugins", priority = 0)]
-        private static void PluginsFolderCreate() => FolderCreate(@"Assets/Plugins", true);
+        [MenuItem(SPECIAL_FOLDERPATH + "Plugins", priority = 15)]
+        private static void PluginsFolderCreate() => CreateDirectoryWithDummyFile(@"Assets/Plugins", true);
         /// <summary> 特別なフォルダを生成します </summary>
-        [MenuItem(SPECIAL_FOLDERPATH + "StreamingAssets", priority = 0)]
-        private static void StreamingAssetsFolderCreate() => FolderCreate(@"Assets/StreamingAssets", true);
+        [MenuItem(SPECIAL_FOLDERPATH + "StreamingAssets", priority = 15)]
+        private static void StreamingAssetsFolderCreate() => CreateDirectoryWithDummyFile(@"Assets/StreamingAssets", true);
 
         #endregion
 
-        private static void FolderCreate(string createDirectory, bool isRefresh = false)
+        /// <summary>
+        /// ディレクトリの生成とともにダミーファイルも作成
+        /// </summary>
+        /// <param name="createDirectory"></param>
+        /// <param name="isRefresh"></param>
+        private static void CreateDirectoryWithDummyFile(string createDirectory, bool isRefresh = false)
         {
             string path = createDirectory;
 
@@ -91,37 +96,42 @@ namespace GDAT.Editor
         }
 
         /// <summary>
-        /// 適切なダミーファイルの使用方法を適用
+        /// 選択されたフォルダ内で適切なダミーファイルの使用方法を適用
         /// フォルダ内にダミーファイル以外がある=消す
         /// フォルダ内にファイルがない=作る
         /// </summary>
-        [MenuItem(MENU_ROOT + "ApplyBestDummyFileUsage", priority = 0)]
+        [MenuItem(ASSETS_MENU_ROOT + "ApplyBestDummyFileUsage", priority = 15)]
         private static void ApplyBestDummyFileUsage()
         {
-            string rootFolderPath = "Assets";
-
-            string[] subdirectories = Directory.GetDirectories(rootFolderPath, "*", SearchOption.AllDirectories);
-
-            foreach (string subdirectory in subdirectories)
+            foreach (var asset in Selection.objects)
             {
-                string dummyFilePath = $@"{subdirectory}/dummy.txt";
+                if (asset is DefaultAsset)
+                {
+                    string rootFolderPath = AssetDatabase.GetAssetPath(asset);
 
-                if (File.Exists(dummyFilePath))
-                {
-                    if (Directory.GetFiles(subdirectory).Length > 2)
+                    string[] subdirectories = Directory.GetDirectories(rootFolderPath, "*", SearchOption.AllDirectories);
+
+                    foreach (string subdirectory in subdirectories)
                     {
-                        UnityEngine.Debug.Log($"{subdirectory}からダミーファイルを削除！");
-                        FileUtil.DeleteFileOrDirectory(dummyFilePath);
+                        string dummyFilePath = $@"{subdirectory}/dummy.txt";
+
+                        if (File.Exists(dummyFilePath))
+                        {
+                            if (Directory.GetFiles(subdirectory).Length > 2)
+                            {
+                                UnityEngine.Debug.Log($"{subdirectory}からダミーファイルを削除！");
+                                FileUtil.DeleteFileOrDirectory(dummyFilePath);
+                            }
+                            continue;
+                        }
+                        if (Directory.GetFiles(subdirectory).Length == 0)
+                        {
+                            UnityEngine.Debug.Log($"{subdirectory}にダミーファイルを生成！");
+                            using (File.Create(dummyFilePath)) { }
+                        }
                     }
-                    continue;
-                }
-                if (Directory.GetFiles(subdirectory).Length == 0)
-                {
-                    UnityEngine.Debug.Log($"{subdirectory}にダミーファイルを生成！");
-                    using (File.Create(dummyFilePath)) { }
                 }
             }
-
             AssetDatabase.Refresh();
         }
     }
